Podstawowa walidacja dodawanego stanowiska wraz z komunikatami do użytkownika o tym co poszło nie tak.

+++ src/main/java/com/WorkersDataBase/service/position/PositionService.java
@@ -1,7 +1,10 @@
 package com.WorkersDataBase.service.position;
 
-import com.WorkersDataBase.data.contract.Position;
-import com.WorkersDataBase.data.contract.PositionRepository;
+import com.WorkersDataBase.data.contract.ContractRepository;
+import com.WorkersDataBase.data.position.Position;
+import com.WorkersDataBase.data.position.PositionRepository;
+import com.WorkersDataBase.data.worker.WorkerRepository;
+import com.WorkersDataBase.service.worker.ServicePushNotification;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,21 +15,44 @@ import java.util.List;
 @AllArgsConstructor
 public class PositionService {
     private final PositionRepository positionRepository;
+    private final ContractRepository contractRepository;
+    private final ServicePushNotification servicePushNotification;
+
     //  addNewPositionType returns true if adding was successful
-    public boolean addNewPositionType(String name){
-        Position position = new Position();
-        position.setPositionName(name);
-        positionRepository.save(position);
-        return true;
+    @Transactional
+    public boolean addNewPositionType(String positionName){
+        boolean positionNameIsUnique = !positionRepository.existsByPositionName(positionName);
+
+        if (positionNameIsUnique){
+            Position position = new Position();
+            position.setPositionName(positionName);
+            positionRepository.save(position);
+
+            servicePushNotification.pushSuccess("Dodano stanowisko " + positionName + ".");
+            return true;
+        }
+
+        servicePushNotification.pushError("Stanowisko powinno mieć unikalną nazwę.");
+        return false;
     }
 
     public List<Position> getPositions(){
         return positionRepository.findAll();
     }
 
+    //  deletePosition returns true if deleting was successful
     @Transactional
-    public void deletePosition(Position position){
+    public boolean deletePosition(Position position){
+        boolean someoneHasContractWithThisPosition =
+                contractRepository.existsByPosition(position);
+
+        if (someoneHasContractWithThisPosition){
+            servicePushNotification.pushError("Ktoś ma to stanowisko na umowie. Najpierw rozwiąż ten problem.");
+            return false;
+        }
 
         positionRepository.delete(position);
+        servicePushNotification.pushSuccess("Usunięto stanowisko " + position.getPositionName() + ".");
+        return true;
     }
 }
